tcapflow-client: Adopt gRPC for client (and later server) handling
This will allow to capture on multiple nodes and then build the
state table on the central one. This adds the client. the gRPC
code was created with:

protoc rpc/tcapcollection.proto --go_out=plugins=grpc:.

gRPC will continously try to connect to the remote system and we
do not retry failed requests. Just log them for now. With statsd
we could see incomplete states relate to number of failed requests.