tcapflow-server: Add a simple collection server
Implement the gRPC server to receive TCAP state. Keep all of it
in RAM. Deal with a multi-node capture setup where the TC-end or
TC-abort or TC-continue arrive at the server before TC-begin. Add
a second list to keep track of dialogues that we stopped to track
but might add furhter messages.